+++ packages/marko-web-identity-x/browser/mixins/global-event-emitter.js
